+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -462,6 +462,7 @@ export default class Meeting extends StatelessWebexPlugin {
   shareStatus: string;
   statsAnalyzer: StatsAnalyzer;
   transcription: Transcription;
+  receiveTranscription: boolean;
   updateMediaConnections: (mediaConnections: any[]) => void;
   endCallInitiateJoinReq: any;
   endJoinReqResp: any;
@@ -1761,8 +1762,8 @@ export default class Meeting extends StatelessWebexPlugin {
       LOCUSINFO.EVENTS.CONTROLS_MEETING_TRANSCRIBE_UPDATED,
       ({caption, transcribing}) => {
         // @ts-ignore - config coming from registerPlugin
-        if (transcribing && this.transcription && this.config.receiveTranscription) {
-          this.receiveTranscription();
+        if (transcribing && !this.transcription && this.config.receiveTranscription) {
+          this.startTranscription();
         } else if (!transcribing && this.transcription) {
           Trigger.trigger(
             this,
@@ -2261,8 +2262,17 @@ export default class Meeting extends StatelessWebexPlugin {
         });
       }
     });
-    this.locusInfo.on(LOCUSINFO.EVENTS.SELF_ADMITTED_GUEST, (payload) => {
+    this.locusInfo.on(LOCUSINFO.EVENTS.SELF_ADMITTED_GUEST, async (payload) => {
       this.stopKeepAlive();
+      // @ts-ignore
+      if (!this.transcription && (this.config.receiveTranscription || this.receiveTranscription)) {
+        if (this.isTranscriptionSupported()) {
+          await this.startTranscription();
+          LoggerProxy.logger.info(
+            'Meeting:index#setUpLocusInfoSelfListener --> enabled to receive transcription for guest user!'
+          );
+        }
+      }
 
       if (payload) {
         Trigger.trigger(
@@ -3725,9 +3735,9 @@ export default class Meeting extends StatelessWebexPlugin {
    * @private
    * @returns {Promise<void>} a promise to open the WebSocket connection
    */
-  private async receiveTranscription() {
+  private async startTranscription() {
     LoggerProxy.logger.info(
-      `Meeting:index#receiveTranscription -->
+      `Meeting:index#startTranscription -->
       Attempting to generate a web socket url.`
     );
 
@@ -3744,7 +3754,7 @@ export default class Meeting extends StatelessWebexPlugin {
       });
 
       LoggerProxy.logger.info(
-        `Meeting:index#receiveTranscription -->
+        `Meeting:index#startTranscription -->
         Generated web socket url succesfully.`
       );
 
@@ -3756,9 +3766,15 @@ export default class Meeting extends StatelessWebexPlugin {
       );
 
       LoggerProxy.logger.info(
-        `Meeting:index#receiveTranscription -->
+        `Meeting:index#startTranscription -->
         opened LLM web socket connection successfully.`
       );
+
+      if (!this.inMeetingActions.isClosedCaptionActive) {
+        LoggerProxy.logger.error(
+          `Meeting:index#receiveTranscription --> Transcription cannot be started until a licensed user enables it`
+        );
+      }
 
       // retrieve and pass the payload
       this.transcription.subscribe((payload) => {
@@ -3777,7 +3793,7 @@ export default class Meeting extends StatelessWebexPlugin {
       // @ts-ignore - fix type
       this.transcription.connect(this.webex.credentials.supertoken.access_token);
     } catch (error) {
-      LoggerProxy.logger.error(`Meeting:index#receiveTranscription --> ${error}`);
+      LoggerProxy.logger.error(`Meeting:index#startTranscription --> ${error}`);
       Metrics.sendBehavioralMetric(BEHAVIORAL_METRICS.RECEIVE_TRANSCRIPTION_FAILURE, {
         correlation_id: this.correlationId,
         reason: error.message,
@@ -3950,6 +3966,7 @@ export default class Meeting extends StatelessWebexPlugin {
       .then((join) => {
         joinSuccess(join);
         this.deferJoin = undefined;
+        this.receiveTranscription = !!options.receiveTranscription;
         Metrics.sendBehavioralMetric(BEHAVIORAL_METRICS.JOIN_SUCCESS, {
           correlation_id: this.correlationId,
         });
@@ -3959,10 +3976,10 @@ export default class Meeting extends StatelessWebexPlugin {
       .then(async (join) => {
         if (isBrowser) {
           // @ts-ignore - config coming from registerPlugin
-          if (this.config.receiveTranscription || options.receiveTranscription) {
+          if (this.config.receiveTranscription || this.receiveTranscription) {
             if (this.isTranscriptionSupported()) {
-              await this.receiveTranscription();
-              LoggerProxy.logger.info('Meeting:index#join --> enabled to recieve transcription!');
+              await this.startTranscription();
+              LoggerProxy.logger.info('Meeting:index#join --> enabled to receive transcription!');
             }
           }
         } else {
@@ -4513,6 +4530,35 @@ export default class Meeting extends StatelessWebexPlugin {
           mediaType: data.type,
         },
       });
+    });
+    this.statsAnalyzer.on(StatsAnalyzerEvents.NO_VIDEO_ENCODED, (data) => {
+      Metrics.sendBehavioralMetric(BEHAVIORAL_METRICS.NO_VIDEO_ENCODED);
+      Trigger.trigger(
+        this,
+        {
+          file: 'meeting/index',
+          function: 'compareLastStatsResult',
+        },
+        EVENT_TRIGGERS.MEETING_NO_VIDEO_ENCODED,
+        data
+      );
+    });
+    this.statsAnalyzer.on(StatsAnalyzerEvents.NO_FRAMES_SENT, (data) => {
+      if (
+        (this.mediaProperties.mediaDirection?.sendVideo && data.mediaType === 'video') ||
+        (this.mediaProperties.mediaDirection?.sendShare && data.mediaType === 'share')
+      ) {
+        Metrics.sendBehavioralMetric(BEHAVIORAL_METRICS.NO_FRAMES_SENT);
+        Trigger.trigger(
+          this,
+          {
+            file: 'meeting/index',
+            function: 'compareLastStatsResult',
+          },
+          EVENT_TRIGGERS.MEETING_NO_FRAMES_SENT,
+          data
+        );
+      }
     });
   };
 
@@ -5698,7 +5744,7 @@ export default class Meeting extends StatelessWebexPlugin {
     if (layoutType) {
       if (!LAYOUT_TYPES.includes(layoutType)) {
         return this.rejectWithErrorLog(
-          'Meeting:index#changeVideoLayout --> cannot change video layout, invalid layoutType recieved.'
+          'Meeting:index#changeVideoLayout --> cannot change video layout, invalid layoutType received.'
         );
       }