+++ packages/@webex/plugin-meetings/src/member/index.ts
@@ -2,6 +2,7 @@
  * Copyright (c) 2015-2020 Cisco Systems, Inc. See LICENSE file.
  */
 import {MEETINGS, _IN_LOBBY_, _NOT_IN_MEETING_, _IN_MEETING_} from '../constants';
+import {IMediaStatus} from './member.types';
 
 import MemberUtil from './util';
 
@@ -10,6 +11,7 @@ import MemberUtil from './util';
  */
 export default class Member {
   associatedUser: any;
+  canReclaimHost: boolean;
   id: any;
   isAudioMuted: any;
   isContentSharing: any;
@@ -28,6 +30,7 @@ export default class Member {
   isSelf: any;
   isUser: any;
   isVideoMuted: any;
+  mediaStatus: IMediaStatus;
   name: any;
   participant: any;
   status: any;
@@ -55,6 +58,13 @@ export default class Member {
         }
       | any = {}
   ) {
+    /**
+     * @instance
+     * @type {Boolean}
+     * @public
+     * @memberof Member
+     */
+    this.canReclaimHost = false;
     /**
      * The server participant object
      * @instance
@@ -220,6 +230,17 @@ export default class Member {
      * @memberof Member
      */
     this.isModeratorAssignmentProhibited = null;
+    /**
+     * @instance
+     * @type {IMediaStatus}
+     * @public
+     * @memberof Member
+     * @example {audio: MediaStatus.RECVONLY, video: MediaStatus.SENDRECV}
+     */
+    this.mediaStatus = {
+      audio: null,
+      video: null,
+    };
     // TODO: more participant types
     // such as native client, web client, is a device, what type of phone, etc
     this.processParticipant(participant);
@@ -237,6 +258,7 @@ export default class Member {
   private processParticipant(participant: object) {
     this.participant = participant;
     if (participant) {
+      this.canReclaimHost = MemberUtil.canReclaimHost(participant);
       this.id = MemberUtil.extractId(participant);
       this.name = MemberUtil.extractName(participant);
       this.isAudioMuted = MemberUtil.isAudioMuted(participant);
@@ -294,6 +316,8 @@ export default class Member {
       this.isAudioMuted,
       this.type
     );
+
+    this.mediaStatus = MemberUtil.extractMediaStatus(this.participant);
   }
 
   /**