+++ packages/@webex/plugin-meetings/src/statsAnalyzer/global.ts
@@ -76,20 +76,11 @@ const STATS_DEFAULT = {
     },
   },
   resolutions: {
-    audio: {
-      send: {
-        width: 0,
-        height: 0,
-      },
-      recv: {
-        width: 0,
-        height: 0,
-      },
-    },
     video: {
       send: {
         width: 0,
         height: 0,
+        framesSent: 0,
       },
       recv: {
         width: 0,
@@ -100,6 +91,7 @@ const STATS_DEFAULT = {
       send: {
         width: 0,
         height: 0,
+        framesSent: 0,
       },
       recv: {
         width: 0,