+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -770,12 +770,12 @@ describe('plugin-meetings', () => {
           assert.equal(meeting.isTranscriptionSupported(), true);
         });
       });
-      describe('#receiveTranscription', () => {
+      describe('#startTranscription', () => {
         it('should invoke subscribe method to invoke the callback', () => {
           meeting.monitorTranscriptionSocketConnection = sinon.stub();
           meeting.initializeTranscription = sinon.stub();
 
-          meeting.receiveTranscription().then(() => {
+          meeting.startTranscription().then(() => {
             assert.equal(true, false);
             assert.calledOnce(meeting.initializeTranscription);
             assert.calledOnce(meeting.monitorTranscriptionSocketConnection);
@@ -786,7 +786,7 @@ describe('plugin-meetings', () => {
           meeting.request = sinon.stub().returns(Promise.reject());
 
           try {
-            await meeting.receiveTranscription();
+            await meeting.startTranscription();
           } catch (err) {
             assert(err, {});
           }
@@ -840,12 +840,12 @@ describe('plugin-meetings', () => {
             assert.calledOnce(MeetingUtil.joinMeeting);
             assert.calledOnce(meeting.setLocus);
           });
-          it('should invoke `receiveTranscription()` if receiveTranscription is set to true', async () => {
+          it('should invoke `startTranscription()` if receiveTranscription is set to true', async () => {
             meeting.isTranscriptionSupported = sinon.stub().returns(true);
-            meeting.receiveTranscription = sinon.stub().returns(Promise.resolve());
+            meeting.startTranscription = sinon.stub().returns(Promise.resolve());
 
             await meeting.join({receiveTranscription: true});
-            assert.calledOnce(meeting.receiveTranscription);
+            assert.calledOnce(meeting.startTranscription);
           });
 
           it('should not create new correlation ID on join immediately after create', async () => {
@@ -1328,6 +1328,71 @@ describe('plugin-meetings', () => {
               event: eventType.MEDIA_QUALITY,
               data: {intervalData: fakeData, networkType: 'wifi'},
             });
+          });
+          it('NO_FRAMES_SENT triggers "meeting:noFramesSent" event and sends metrics', async () => {
+            meeting.mediaProperties.mediaDirection = {sendVideo: true};
+            statsAnalyzerStub.emit(
+              {file: 'test', function: 'test'},
+              StatsAnalyzerModule.EVENTS.NO_FRAMES_SENT,
+              {mediaType: 'video'}
+            );
+
+            assert.calledWith(
+              TriggerProxy.trigger,
+              sinon.match.instanceOf(Meeting),
+              {
+                file: 'meeting/index',
+                function: 'compareLastStatsResult',
+              },
+              EVENT_TRIGGERS.MEETING_NO_FRAMES_SENT,
+              {
+                mediaType: 'video',
+              }
+            );
+            assert.calledWith(Metrics.sendBehavioralMetric, BEHAVIORAL_METRICS.NO_FRAMES_SENT);
+          });
+          it('NO_FRAMES_SENT triggers "meeting:noFramesSent" event and sends metrics for share', async () => {
+            meeting.mediaProperties.mediaDirection = {sendShare: true};
+            statsAnalyzerStub.emit(
+              {file: 'test', function: 'test'},
+              StatsAnalyzerModule.EVENTS.NO_FRAMES_SENT,
+              {mediaType: 'share'}
+            );
+
+            assert.calledWith(
+              TriggerProxy.trigger,
+              sinon.match.instanceOf(Meeting),
+              {
+                file: 'meeting/index',
+                function: 'compareLastStatsResult',
+              },
+              EVENT_TRIGGERS.MEETING_NO_FRAMES_SENT,
+              {
+                mediaType: 'share',
+              }
+            );
+            assert.calledWith(Metrics.sendBehavioralMetric, BEHAVIORAL_METRICS.NO_FRAMES_SENT);
+          });
+          it('NO_VIDEO_ENCODED triggers "meeting:noVideoEncoded" event and sends metrics', async () => {
+            statsAnalyzerStub.emit(
+              {file: 'test', function: 'test'},
+              StatsAnalyzerModule.EVENTS.NO_VIDEO_ENCODED,
+              {mediaType: 'video'}
+            );
+
+            assert.calledWith(
+              TriggerProxy.trigger,
+              sinon.match.instanceOf(Meeting),
+              {
+                file: 'meeting/index',
+                function: 'compareLastStatsResult',
+              },
+              EVENT_TRIGGERS.MEETING_NO_VIDEO_ENCODED,
+              {
+                mediaType: 'video',
+              }
+            );
+            assert.calledWith(Metrics.sendBehavioralMetric, BEHAVIORAL_METRICS.NO_VIDEO_ENCODED);
           });
         });
       });
@@ -3578,7 +3643,48 @@ describe('plugin-meetings', () => {
           );
           done();
         });
+        it('transcription should start when configured when guest admitted', (done) => {
+          meeting.isTranscriptionSupported = sinon.stub().returns(true);
+          meeting.receiveTranscription = sinon.stub().returns(true);
+          meeting.startTranscription = sinon.stub();
+
+          meeting.locusInfo.emit({function: 'test', file: 'test'}, 'SELF_ADMITTED_GUEST', test1);
+          assert.calledOnce(meeting.startTranscription);
+          done();
+        });
       });
+
+      describe('#setupLocusControlsListener', () => {
+        it('transcription should start when meeting transcribe state is updated with active transcribing', (done) => {
+          const payload = {caption: true, transcribing: true};
+          meeting.startTranscription = sinon.stub();
+          meeting.config.receiveTranscription = true;
+          meeting.transcription = null;
+
+          meeting.locusInfo.emit({function: 'meeting/index', file: 'setupLocusControlsListener'}, 'CONTROLS_MEETING_TRANSCRIBE_UPDATED', payload);
+          assert.calledOnce(meeting.startTranscription);
+          done();
+        })
+
+        it('transcription should stop when meeting transcribe state is updated with inactive transcribing', (done) => {
+          const payload = {caption: false, transcribing: false};
+          meeting.startTranscription = sinon.stub();
+          meeting.config.receiveTranscription = true;
+          meeting.transcription = {};
+
+          meeting.locusInfo.emit({function: 'meeting/index', file: 'setupLocusControlsListener'}, 'CONTROLS_MEETING_TRANSCRIBE_UPDATED', payload);
+          assert.notCalled(meeting.startTranscription);
+          assert.calledTwice(TriggerProxy.trigger);
+          assert.calledWith(
+            TriggerProxy.trigger,
+            sinon.match.instanceOf(Meeting),
+            {file: 'meeting/index', function: 'setupLocusControlsListener'},
+            'meeting:receiveTranscription:stopped',
+            payload
+          );
+          done();
+        })
+      })
 
       describe('#setUpLocusUrlListener', () => {
         it('listens to the locus url update event', (done) => {
@@ -3605,14 +3711,17 @@ describe('plugin-meetings', () => {
       describe('#setUpLocusServicesListener', () => {
         it('listens to the locus services update event', (done) => {
           const newLocusServices = {
-              services: {
-                record: {
-                  url: 'url',
-                }
+            services: {
+              record: {
+                url: 'url',
               },
+            },
           };
 
-          meeting.recordingController = {setServiceUrl: sinon.stub().returns(undefined), setSessionId: sinon.stub().returns(undefined)};
+          meeting.recordingController = {
+            setServiceUrl: sinon.stub().returns(undefined),
+            setSessionId: sinon.stub().returns(undefined),
+          };
 
           meeting.locusInfo.emit(
             {function: 'test', file: 'test'},
@@ -3620,7 +3729,10 @@ describe('plugin-meetings', () => {
             newLocusServices
           );
 
-          assert.calledWith(meeting.recordingController.setServiceUrl, newLocusServices.services.record.url);
+          assert.calledWith(
+            meeting.recordingController.setServiceUrl,
+            newLocusServices.services.record.url
+          );
           assert.calledOnce(meeting.recordingController.setSessionId);
           done();
         });