+++ packages/@webex/plugin-meetings/test/unit/spec/member/index.js
@@ -5,18 +5,50 @@ import MemberUtil from '@webex/plugin-meetings/src/member/util';
 import Member from '@webex/plugin-meetings/src/member';
 
 describe('member', () => {
+  const participant = {controls: {}, status: {}};
+
+  const member = new Member(participant);
+
   afterEach(() => {
     sinon.restore();
   });
 
-  it('checks that processParticipant calls isHandRaised', () => {
-    const participant = {controls: {}};
+  describe('#processParticipant', ()=>{
+    it('checks that processParticipant calls isHandRaised', () => {
+      sinon.spy(MemberUtil, 'isHandRaised');
+      member.processParticipant(participant);
+  
+      assert.calledOnceWithExactly(MemberUtil.isHandRaised, participant);
+    });
 
-    const member = new Member({});
+    it('checks that processParticipant calls canReclaimHost', () => {
+      sinon.spy(MemberUtil, 'canReclaimHost');
+      member.processParticipant(participant);
 
-    sinon.spy(MemberUtil, 'isHandRaised');
-    member.processParticipant(participant);
+      assert.calledOnceWithExactly(MemberUtil.canReclaimHost, participant);
+    });
+  })
 
-    assert.calledOnceWithExactly(MemberUtil.isHandRaised, participant);
-  });
+  describe('#processMember', ()=>{
+    it('checks that processMember calls isRemovable', () => {
+      sinon.spy(MemberUtil, 'isRemovable');
+      member.processMember();
+  
+      assert.calledOnce(MemberUtil.isRemovable);
+    });
+  
+    it('checks that processMember calls isMutable', () => {
+      sinon.spy(MemberUtil, 'isMutable');
+      member.processMember();
+  
+      assert.calledOnce(MemberUtil.isMutable);
+    });
+  
+    it('checks that processMember calls extractMediaStatus', () => {
+      sinon.spy(MemberUtil, 'extractMediaStatus');
+      member.processMember();
+  
+      assert.calledOnceWithExactly(MemberUtil.extractMediaStatus, participant);
+    });
+  })
 });