+++ packages/@webex/plugin-meetings/test/unit/spec/member/util.js
@@ -47,3 +47,66 @@ describe('isHandRaised', () => {
     assert.isFalse(MemberUtil.isHandRaised(participant));
   });
 });
+
+describe('extractMediaStatus', () => {
+  it('throws error when there is no participant', () => {
+    assert.throws(() => {
+      MemberUtil.extractMediaStatus()
+    }, 'Media status could not be extracted, participant is undefined.');
+  });
+
+  it('returns undefined media status when participant audio/video status is not present', () => {
+    const participant = {
+      status: {}
+    };
+    
+    const mediaStatus = MemberUtil.extractMediaStatus(participant)
+
+    assert.deepEqual(mediaStatus, {audio: undefined, video: undefined});
+  });
+
+  it('returns correct media status when participant audio/video status is present', () => {
+    const participant = {
+      status: {
+        audioStatus: 'RECVONLY',
+        videoStatus: 'SENDRECV'
+      }
+    };
+    
+    const mediaStatus = MemberUtil.extractMediaStatus(participant)
+
+    assert.deepEqual(mediaStatus, {audio: 'RECVONLY', video: 'SENDRECV'});
+  });
+});
+
+describe('MemberUtil.canReclaimHost', () => {
+  it('throws error when there is no participant', () => {
+    assert.throws(() => {
+      MemberUtil.canReclaimHost();
+    }, 'canReclaimHostRole could not be processed, participant is undefined.');
+  });
+
+  it('returns true when canReclaimHostRole is true', () => {
+    const participant = {
+      canReclaimHostRole: true,
+    };
+
+    assert.isTrue(MemberUtil.canReclaimHost(participant));
+  });
+
+  it('returns false when canReclaimHostRole is false', () => {
+    const participant = {
+      canReclaimHostRole: false,
+    };
+
+    assert.isFalse(MemberUtil.canReclaimHost(participant));
+  });
+
+  it('returns false when canReclaimHostRole is falsy', () => {
+    const participant = {
+      canReclaimHostRole: undefined,
+    };
+
+    assert.isFalse(MemberUtil.canReclaimHost(participant));
+  });
+});